fixed the handling ampersands in the comment
The memo field is handled in a weird way: it's an HTML-encoded
XML snippet, which is not necessarily valid XML as some special chars
are not escaped properly. As a simple and ugly fix let's just correct
the far most common such character in real-life note fields for bank
transfers, the ampersand (&). Yep, it's ugly, but this whole parsing
is just a heuristic piece of mess anyway.